# prov/sockets: Optimize startup