These changes improve the `MPI_init`-`MPI_Finalize` test:
BDW+EDR (number of ranks - 176, ppn 1):
- 1.51833218 sec - w/o optimizations (only the sockets provider is enabled)
- 1.4842533 sec - w/ optimizations (only the sockets provider is enabled)
- 1.81241148 sec - w/o optimizations (all available providers is enabled)
- 1.80456235 sec - w/ optimizations (all available providers is enabled)

KNL+OPA (number of ranks -136, ppn 68)
- 11.77717361 sec - w/o optimizations (only the sockets provider is enabled)
- 11.5522065 sec - w/ optimizations (only the sockets provider is enabled)
- 10.33442811 sec - w/o optimizations (all available providers is enabled)
- 10.31751561 sec - w/ optimizations (all available providers is enabled)

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>